Fix conditional rendering of data table!!
If we had conditional rendering of the component, then it wouldn't be rendered with null data initially. However, we were only checking for new data on `componentWillReceiveProps()`. However, this doesn't get called on first render, so we weren't actually calling it until it was rendered for a second time (when it got a SECOND csv?)! SO, instead we're calling `hashArray()` on `componentWillMount()` when it mounts the first time AND `componentWillReceiveProps()`, before it updates each subsequent time.